Move MutableBundleManifest into the tycho-model module for reuse
(cherry picked from commit 46d4640e051175a8a6f3d7fb5ccadd6797624cf2)
