+++ tycho-versions-plugin/src/main/java/org/eclipse/tycho/versions/engine/VersionUpdater.java
@@ -30,8 +30,8 @@ import org.eclipse.tycho.TychoConstants;
 import org.eclipse.tycho.model.Feature;
 import org.eclipse.tycho.model.IU;
 import org.eclipse.tycho.model.ProductConfiguration;
+import org.eclipse.tycho.model.manifest.MutableBundleManifest;
 import org.eclipse.tycho.versions.bundle.MutableBndFile;
-import org.eclipse.tycho.versions.bundle.MutableBundleManifest;
 import org.eclipse.tycho.versions.pom.PomFile;
 import org.eclipse.tycho.versions.utils.ProductFileFilter;
 import org.osgi.framework.Constants;
